code_mvp/requirements.txt
h5py==3.10.0
matplotlib==3.7.5
munch==4.0.0
numpy==1.24.3
open3d==0.18.0
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.3.2
setuptools==68.2.2
torch==1.12.1
tqdm==4.66.2
